+++ CHANGES.md
@@ -16,6 +16,9 @@ Fast and flexible Python library for text tables.
 
 * dev
   * Added option to control column's title visibility.
+  * Added `set_no_data_label()` to override default `NO DATA` shown when rendering empty table.
+  * Fixed rendering of middle row separator when 'NO DATA' label is used
+
 
 * 2.6.0 (2023-05-05)
   * Fixed column container not updating width of columns correctly.
@@ -24,6 +27,7 @@ Fast and flexible Python library for text tables.
   * Added more tests.
   * Updated docs.
   * `FlexTable` c'tor now also accepts rows for instant initialization.
+
 
 * 2.5.0 (2023-04-11)
   * Initial public release.